requirements.txt
aiofiles==23.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.9.1
aiosignal==1.3.2
anyio==4.9.0
appdirs==1.4.4
APScheduler==3.10.1
astroid==3.3.10
attrs==25.3.0
cachetools==4.2.2
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
dill==0.4.0
distlib==0.3.9
filelock==3.18.0
flake8==7.3.0
frozenlist==1.7.0
h11==0.14.0
httpcore==0.16.3
httpx==0.24.1
idna==3.10
isort==6.0.1
mccabe==0.7.0
multidict==6.5.0
openai==0.28.0
platformdirs==4.3.8
propcache==0.3.2
pycodestyle==2.14.0
pyflakes==3.4.0
pylint==3.3.7
PyPDF2==3.0.1
pyTelegramBotAPI==4.27.0
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2025.2
requests==2.31.0
rfc3986==1.5.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
telegram-send==0.37
tomlkit==0.13.3
tornado==6.1
tqdm==4.67.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.5.0
virtualenv==20.31.2
wheel==0.45.1
yarl==1.20.1
